+++ plugins/sdr_sources/airspy_sdr_support/airspy_sdr.cpp
@@ -195,7 +195,7 @@ void AirspySource::drawControlUI()
 
 void AirspySource::set_samplerate(uint64_t samplerate)
 {
-    for (int i = 0; i < available_samplerates.size(); i++)
+    for (int i = 0; i < (int)available_samplerates.size(); i++)
     {
         if (samplerate == available_samplerates[i])
         {
