+++ plugins/sdr_sources/rtlsdr_sdr_support/rtlsdr_sdr.cpp
@@ -141,7 +141,7 @@ void RtlSdrSource::drawControlUI()
 
 void RtlSdrSource::set_samplerate(uint64_t samplerate)
 {
-    for (int i = 0; i < available_samplerates.size(); i++)
+    for (int i = 0; i < (int)available_samplerates.size(); i++)
     {
         if (samplerate == available_samplerates[i])
         {
